[DEV-3997] Reorganized tests a bit
